Standardize versioning in dependabot common
What:
Move most versioning code into depenabot common so
we can handle versioning in a standardized way.

Why:
To reduce issues relating to validity and
comparisons of versions.
